Minimal parameter objects implementation
This adds support for parameter objects to dig. Structs opt-in to be
treated as parameter objects by embedding `dig.Param`.

For the sake of simplicity from the users' point-of-view, we don't
recurse into parameter objects inside other parameter objects. This can
be changed in the future if there's demand for it.