dig.Param: Recurse into nested parameter objects (#73)
This change adds support to parameter objects for recursing into fields that
are also parameter objects. That is, this change allows the following,

    type OutputConfig struct {
        dig.Param

        Writer io.Writer
    }

    type Config struct {
        dig.Param

        Input io.Reader
        Output OutputConfig
    }

    func NewTransport(cfg Config) (*Transport, error) { .. }

Besides the usefulness of treating fields similar to constructor parameters,
there's another motivating use case for this: Aliasing `dig.Param`. Without
this change, we have special handling for the `dig.Param` type, opting it out
of resolution from `dig`. If a library author wants to use their own type in
place of `dig.Param` (for example, `fx.Param`), they can't quite do this
because any alias of `dig.Param` is a different type; one which the library
does not special-case.

    package fx

    type Param dig.Param

    type Config struct {
        Param

        Input io.Reader
    }

    // The above with fail because dig doesn't know how to resolve fx.Param.

With this change, Fx will be able to support the following.

    package fx

    type Param struct{ dig.Param }

    type Config struct {
        Param

        // ..
    }