optional: Fix support for constructors (#94)
The current implementation of optionals has a bug which causes failures
if a constructor for a provided type has `dig.In` struct with optional
parameters.

That is, given,

    type param struct {
        In

        Buffer *bytes.Buffer `optional:"true"`
    }

    c.Provide(func(p param) *Something { ... })

The following will fail, complaining that `*bytes.Buffer` isn't
available in the container.

    c.Invoke(func(*Something) { .. })

This fixes this issue by storing more information about node
dependencies: specifically, whether they are required or optional.

Test Plan: A failing test case was added before the fix was implemented.