Invoke: Return errors as-is (#96)
This changes Invoke to return the error returned by the function as-is
rather than wrapping it in Errorf. As a caller of Invoke, I would expect
my function's returned error to be returned back to me. Otherwise, I'm
losing a lot of information if my error object was richer than
`fmt.Errorf`'s string.

Further, the existing message didn't really add a lot of information: it
added the address of the function to the error message.

Depends on #95 so that tests don't fail.