Manual sentinel value detection (#113)
This changes our detection of `dig.In` and `dig.Out` structs from
relying on Go's automatic wrapper method generation to a manual check
for that embedded value. We do this by simply performing a breadth first
search of embedded fields of a struct.

Note about performance: There usually won't be a lot of embedding in
constructor struct types so the function should return very quickly. If
performance ever becomes an issue, we can memoize types that are known
to embed `dig.In` or `dig.Out` in a global cache similar to how
`encoding/json` caches reflection information for types.

By not relying on wrapper method generation, we are able to support
struct types generated with `reflect.StructOf`, as a documented
limitation of these structs is,

> StructOf currently does not generate wrapper methods for embedded
> fields. This limitation may be lifted in a future version.

This opens up more complex use cases where reflection is required to
generate the `Invoke`d or `Provide`d function.

---

Also note that I added tip to .travis.yml so that we can verify this
behavior with Go 1.9 and newer.