Add typealiases for Invoke and Provide parameters.
This is going to allow us to better document these concepts with
examples. I think the biggest benefit of these is going to come from
migrating `fx` to use these aliases in it's Provide and Invoke functions
meaning the developers would be able to access the docs much easier.
Right now we just say "go find dig docs".