Rewrite godocs
This rewrites most of the godocs of this library and documents
undocumented features (dig.In, dig.Out, etc.).

The package-level docs give a general overview of the library and
showcase different dig.In/dig.Out features by use case.

Entity-level docs for dig.In and dig.Out act as the authoritative
reference list of all tags supported by those structs and how they
behave.