Support disabling error wrapping
This adds support for disabling wrapping of errors returned by
user-provided constructors.

    c := dig.New(dig.WrapErrors(false))

If set to false, this returns errors returned by `Provide`d constructors
as-is, without any additional wrapping.

Implementation:
We use a custom error wrapper for `fmt.Errorf` calls which include the
original error at the end. So,

    return fmt.Errorf("something went wrong: %v", err)

Becomes,

    return errWrapf(err, "something went wrong")

The wrapper tracks the original `err` and makes it available later.

The error messages remain the same but if wrapping is disabled, we will
return the original error instead of the wrapped one from the top-most
layer (the `Invoke` call).

Resolves #145