Better error messaging during mixed types usage
This is something that has been mentioned several times - it's very easy
to request a wrong type on accident - is it a pointer? is it value type?
In a large enough framework base it's a giant mix of both. This adds a
much better error message when dig is able to detect an easily
correctable case.

Fixes #147