Add RootCause to retrieve original error (#146)
This adds a new `dig.RootCause(error) error` function which retrieves the
original error that caused an Invoke to fail.

See #145 for motivation but the gist of it is, this makes it possible to
retrieve and act upon the original error that a constructor of a dependency
returned when an Invoke call is made.

Implementation:
We use a custom error wrapper for `fmt.Errorf` calls which include the
original error at the end. So,

    return fmt.Errorf("something went wrong: %v", err)

Becomes,

    return errWrapf(err, "something went wrong")

The wrapper tracks the original `err` and `RootCause` is just an accessor for
it.

Because this came up offline: We're not using pkg/errors because we want to
return the error returned by the user-provided constructor. If we use pkg/
errors, and the user-provided constructor also uses pkg/errors, we won't
return what the user returned but rather, the root cause of that error.

Resolves #145