Provide: Inspect constructor parameters upfront
As a first-step to DRYing up the graph resolution logic in dig, this
changes Provide to inspect constructors in more detail upfront.

Their dependencies, whether they are dig.In objects, etc. is now
known on the Provide path rather than Invoke.

This information is stored with the help of a `param` interface which
unionizes the following parameter types:

- paramList: List of parameters of a constructor
- paramSingle: Basic argument type. Nothing special about this
- paramObject: `dig.In` object whose fields are other `param`s

This opens room for more `param` types (tagged types?).

Note: The `Dependencies()` method will be removed in a future change. It
exists only to minimize the impact of this specific change.