Shallow dependency check for `param` types
This implements a shallow check (previously implemented by
`Container.contains`) in terms of `param`. This check just verifies that
constructors for all dependencies of the provided param are present in
the container.

Note that this is the last usage of `Dependencies()` and we should be
able to delete it now.