param: Rename placeholder method to paramType
This renames the placeholder method for the `param` interface to
`paramType`. Otherwise, it's impossible to create an invalid instance of
the interface with use of embedding because the field and method name
conflict.

    type badParam struct{ param }

We can't do the above without the rename because the field `param`
conflicts with the inherited method `param()`.