Feed results into container using `result`s
This changes how values are fed into the container to rely on individual
`result` implementations.

Instead of a single place where we handle the different return values of
a constructor, each `result` type now extracts information it cares
about.

- resultList feeds individual values into the underlying results
- resultSingle stores the value into the container cache
- resultError retrieves and returns the error as-is
- resultObject extracts corresponding dig.Out fields and delegates to
  the underlying results

With this change, both the input and output paths have been refactored
to use this approach.