Provide: Inspect constructor parameters upfront (#149)
As a first-step to DRYing up the graph resolution logic in dig, this
changes Provide to inspect constructors in more detail upfront.

Their dependencies, whether they are dig.In objects, etc. is now
known on the Provide path rather than Invoke.

This information is stored with the help of a `param` interface which
unionizes the following parameter types:

- paramList: List of parameters of a constructor
- paramSingle: Basic argument type. Nothing special about this
- paramObject: `dig.In` object whose fields are other `param`s

This opens room for more `param` types.

Note: The `Dependencies()` method will be removed in a future change. It
exists only to minimize the impact of this specific change.

Also note that no new errors are being surfaced to the user. All errors
returned here are copies of existing errors returned by the library. Of
note is the logic in `newParamObject` which returns an error when
private fields are encountered. In dig currently, that logic exists in
two places:

- https://github.com/uber-go/dig/blob/809ed57c66b4c4b74e882be9da77d997ec439cde/dig.go#L363-L367
- https://github.com/uber-go/dig/blob/809ed57c66b4c4b74e882be9da77d997ec439cde/dig.go#L532-L534

If you look at the implementations, one of them skips private fields
while the other returns an error. The one that errors out does so at
`Invoke` time rather than `Provide`. With this change, the error will be
returned at `Provide` time instead, which is preferable anyway. This is
not actually a breaking change because we're failing fast rather than
waiting for an Invoke.