Use result.Produces() to inject nodes (#157)
This changes the Provide path to rely on `result.Produces()` to inject
nodes into the graph.

This has a significant internal behavioral change: Instead of creating a
new `node` for each type in the container, we now create a new `node`
for each constructor in the container and link all keys produced by that
constructor to the node.

Given the constructor,

    func NewThings(*Dep1, *Dep2) (*Thing1, *Thing2)

We previously had,

    node{Thing1, NewThings}  node{Thing2, NewThings}
                   |                        |
                   |        +---------------+
                   |        |
            Dep1 <-+--------+-> Dep2

Where each node inspected the signature of `NewThings` by itself. With
this change, the graph looks similar to,

    Thing1      Thing2
      |          |
      v          v
     node{NewThings}
      |          |
      v          v
    Dep1        Dep2

Not hugely relevant to this library, but this should also result in a
slight performance gain because we're not inspecting each constructor's
parameters multiple times.