param: Replace forEachParamSingle with visitor
This drops forEachParamSingle in favor of a go/ast-style visitor. The
visitor is called on a param, and if it returns a non-nil visitor, the
returned visitor is called on the children of that param.

This allows us to replace param switches which just want to traverse the
tree. This also lets us ensure that we have only one place in the code
where we panic with "unknown param type".