Constructor error should not cause bad state (#166)
This fixes a regression introduced by refactoring that left the
container in a bad state if a constructor failed.

The issue is that in the following,

    c.Invoke(func(myDepedency) { ... })
    c.Invoke(func(myDepedency) { ... })

If the constructor for myDepedency always fails, the function will still
be called on the second invoke because we will have recorded the output
of the myDepedency constructor from the first call.

This happens because `result.Extract` feeds its results directly into
the container. To fix this issue, I changed result.Extract to operate on
a resultReceiver interface instead of operating directly on the
container. An implementation of resultReceiver stages the pending
changes from a call and commits them to the container only if no errors
were encountered.

This also lets us change result.Extract to no longer return an error
because the error is sent directly to the resultReceiver.