param: Simplify build errors
Currently the errors for param.Build failure add a piece of information
at every level.

    failed to get arguments for constructor ..:
    failde to build parameter 1:
    failed to build object Foo:
    failed to build field bar:
    failed to build type Baz:
    [..]

Everything between the constructor and the actual type is redundant
information. The dependencies of a constructor are always obvious and
we don't need to amend this information to the error message. This
changes the output to,

    failed to build arguments for constructor ...:
    failed to build type baz:
    [..]