node.Call: Always check dependencies first
This changes node.Call to always check for dependencies before trying to
build. The result of this is that *all* constructor calls, not just
those for paramSingle, will complain about missing dependencies early.

Plus when we complain about dependencies, we always talk about the
function intsead of the type now.

This gives us error messages similar to the following for all
constructor calls (including value group callS).

    could not build arguments for <some function>:
    failed to build <type1>:
    missing dependencies for <function producing type1>:
    type <type1> is not in the container, did you mean to provide it?