Provide: Add Name option
This adds a `Name(string)` option to `Provide()` so that the following,

    c.Provide(NewThing, Name("foo"))

Is the same as the following,

    type result struct {
        dig.Out

        Thing *Thing `name:"foo"`
    }

    c.Provide(func(...) result {
        return result{Thing: NewThing(..)}
    })

The name applies to *all* values produced by the constructor--even if it
produces multiple values. If the constructor returns a Result Object,
the name is applied to all fields of that Result object that don't
already have a name.

    type result struct {
        dig.Out

        Foo Foo `name:"foo"`
        Bar Bar
    }

    c.Provide(func(...) result {
        ...
    }, Name("bar"))

    // Name("bar") applies only to field Bar because Foo already has a
    // name.

This change was motivated by an offline discussion with @bayesianmind.
We agreed that Fx needs to support some way to specify the name for a
value produced by a constructor without rewriting the constructor to
produce an fx.Out. An API similar to the following was suggested.

    fx.Provide(fx.Named("foo", NewThing))

Regardless of the user-facing API, such functionality would need to rely
on reflection pretty heavily. It would need to generate a function using
`reflect.MakeFunc` equivalent to the following,

    func(...) (out struct {
        fx.Out

        Result *Thing `name:"foo"`
    }) {
        out.Result = NewThing(...)
        return
    }

See yarpc/yarpc-go#1200 for an example of the initial attempt made by
@bayesianmind.

I don't think that this functionality should rely this heavily on some
fairly fragile reflection logic. The way to build anonymous structs with
embedded fields is not the same between Go 1.8 and 1.9 (see
https://github.com/uber-go/dig/blob/a84ae412585ef6c4b8fc2f7577ff80f701317314/dig_test.go#L628-L635).

Supporting this in Dig natively is much more straightforward and safer,
and it requires no reflection.

A good/bad side-effect of this change is that `name:"..."` tags on
nested Result Objects are now treated as if the `Name(..)` option was
passed for that field. so if we do,

    // package thing

    type Result struct {
        Out

        Thing *Thing
    }

    // package somemodule

    type Result struct {
        Out

        ThingResult thing.Result `name:"foo"`
        AnotherThing *Thing
    }

The value under `somemodule.ThingResult.Thing` will have the name "foo".