dig.Name: Validate that names are consumable
For names to be consumable from a dig.In, they must be strings that can
be represented inside a Go backquote string literal. The only limitation
this has is that the string cannot contain a backquote
(https://golang.org/ref/spec#raw_string_lit).

This commit adds validation for names provided via dig.Name.