result: Remove resultError type
This changes how error return values are represented by dig. Previously,
each return value had a corresponding `result` object. This complicated
errors a bit because they can only exist as positional return values of
a constructor.

This change drops this complexity by making it so that errors are only
tracked at the `resultList` (which represents one or more return values
of a constructor) level.

This change was motivated by an upcoming change which is attempting to
reduce the number of places in which dig abstractions access private
unexported details of each other.