internal: Separate Key types
Previously, we had a single `key` type,

    type key struct {
        t     reflect.Type
        name  string
        group string
    }

This was used in two ways:

- Simple value: Just `t` and `name` set
- Grouped value: Just `t` and `group` set

This change breaks `key` into separate internal `ValueKey` and
`GroupKey` types and unifies them with a `Key` interface.

We can now get a type-level guarantee of whether a key is meant to be
used for a simple value or a value group.