result: Track produced keys upon creation
This changes `result` implementations to track the `internal.Key`s
produced by them at creation rather than walking the tree afterwards.

To keep conflict detection DRY, we introduce an `internal.KeySet` which
is an ordered collection of keys. Duplicates with the same name and type
are not allowed if they are ValueKeys.

This allows us to drop the overly complex connectionVisitor business and
the related resultVisitor interface.