dig.Out: Nested result structs must not be named
This changes the system to disallow specifying names using either
`dig.Name` or `name:".."` tags on result structs.

So users will be unable to do both,

    type Result struct {
        dig.Out

        ...
    }

    c.Provide(func(...) Result {
        ...
    }, dig.Name("foo"))

And,

    type Result struct {
        dig.Out

        Foo struct {
            dig.Out

            Thing *Thing
        } `name:"foo"`
    }

Note that previously, the second case (`name:".."` tag on an embedded
dig.Out) was a no-op. Dig would ignore that tag.

With this change, we are explicitly disallowing this with an error. This
is not a breaking change because that was unsupported/invalid usage.

This change makes room for a backwards-compatible change in the future
which decides how naming on nested result structs will actually work.