dot Group: Add Group nodes to graph (#194)
Group is introduced to represent value groups. It will be a special node in the graph. Constructors that depend on or provide to a value group will have edges entering or leaving the group node. This separates them from regular params and reduces the number of edges in the graph.