detectCycles: Avoid building keys twice
Since we're already type-casting the param to the leaf types to build
the key and get their providers, this simplifies the code a bit further
by using the existing `key` objecst and avoiding the type cast.

We end up with a small amount of duplication in the code. We can avoid
it by doing the check outside the `switch` block but that means we'll
look up the providers for a type even if we don't need them. In this
case, the slight performance win might be worth the extra three lines.

This also makes the keyer interface and the related changes there
unnecessary.

Finally, if it matters: Getting rid of double-casting had a *slight*
performance win.

Before:

    BenchmarkProvideCycleDetection-4            5000            339305 ns/op

After:

    BenchmarkProvideCycleDetection-4            5000            323540 ns/op