walkParam: Don't exit early on nil
Returning `nil` from a visitor only indicates that we don't want to
recurse further into that parameter. We shouldn't use that to exit the
rest of the iteration.

Given,

    type Params struct {
        dig.In

        ...
    }

    func(io.Reader, Params, io.Writer)

If our visitor returns nil on Params because it doesn't care about
parameter objects, we should still look at io.Writer.

This adds a test for this behavior as well.