tests: Better use of assertErrorMatches
The test for DeferAcyclicVerification was making N separate calls to
assertErrorMatches which doesn't verify that the different components of
the error messages appear in the correct order.

This switches to "correct" usage so that we're verifying that the
correct types appear in the error messages at the correct spot. In this
case, we want "C depends on B depends on A depends on C".