Prune graph entries that do not contain failing results
Added a Prune method to remove entries from the graph that do not
contain any failing results. Graph nodes that are not root or
transitive failures do not help the user debug and make it
more difficult to debug by cluttering the graph.

Orient the graph left to right for graph readability