result: Merge resultAs and resultSingle
The dig.As functionality shouldn't introduce its own result type because
this isn't a new leaf kind. It's another option for the case when a
single result is being provided.

This change merges resultSingle and resultAs. In the process, the
conversion of the dig.As arguments to reflect.Types is done once and
re-used throughout.