errMissing[Many]Type[s]: Merge and Format
errMissingType and errMissingManyTypes are errors returned when one or
more types are missing in the container respectively. Differentiating
the two types is unnecessary since the logic is largely the same.

As part of implementing fmt.Formatter for these types, this change
merges the two types into errMissingTypes.

As with prior fmt.Formatter implementations, this behaves differently
based on `%v` or `%+v`.

With `%v`, the error will be printed on a single line as before.

    missing types: foo; bar (did you mean *bar); baz

With `%+v`, the error will be printed across multiple lines as an
itemized list.

    missing types:
    	- foo
    	- bar (did you mean to use *bar)
    	- baz