error/test: assertErrorMatches: Verify %v and %+v
This changes assertErrorMatches to verify the output for both, the `%v`
and the `%+v` versions of an error.

For the `%+v` version, we require that there's a newline between each
line in `msgs` so that where you break the string is no longer a
readability concern but something to be tested. This caused some tests
to be modified slightly.

For example, if you have an error that takes the following form in
`%+v`.

    could not process entry: foo

This will no longer match with the following.

    assertErrorMatches(t, err, "could not process:", "foo")

Instead, the string must either be joined,

    assertErrorMatches(t, err, "could not process: foo")

Or the error be changed to be split across multiple lines in case of
`%+v`.

    could not process entry:
    foo

(The function had to be changed to use a `*testing.T` instead of
`testing.TB` because the latter doesn't support subtests.)