internal/graph: Don't backtrack
We don't need to backtrack to figure out how we got to a cycle.
We can keep track as we recurse.

This helps simplify the implementation a bit,
and obviates the need to reverse the computed path later.

This also gives us the full path we took to find the cycle,
but I pruned it down to just the cycle to match prior behavior.

Finally, I changed the tests to assert the exact/full path of the cycle,
which required updating the test cases a little because they were not
previously validating that.