graphHolder: Don't record orders in a map
Recording orders in a map keyed by `key` is a bit risky.
We'll have to change that as soon as we start using this
reprsentation for more than cycle detection.

Instead, make every node responsible for knowing its order.
As long as we have a means of referencing the node, we know its order.

Specifically, the following changes were made:

- newGraphNode: Drop the `key` parameter and return the order of the newly
  added node instead.
    - newConstructorNode: Record the order returned by newGraphNode on the
      node.
    - newParamGroupedSlice: Record the order returned by newGraphNode on the
      node.
- newGraphNode: Move to containerStore as this functionality is not
  needed in the containerWriter. This lets us delete the newGraphNode
  method from stagingContainerWriter, which we use only at invoke time
  to stage changes to the container.
- provider: Add an `Order() int` accessor to report the position of that
  provider in graphHolder.
- graphHolder: delete orders map, return orders on add instead.
- graphHolder: Simplify snapshot to just dropping the newly added nodes.