Introduce Scope to Dig (#305)
This introduces the concept of Dig.Scope, which is a scoped container for modifications to the dependency graph. A Scope may have one or more Scopes as children. A Scope also has one parent Scope, except for Container, which is the root Scope. This change is necessary before we can start working on APIs for graph modifications and decorations.

Things to note:

Modifications to a Scope is "private" by default. For example, Provide to a Scope means it will be made available to itself and any of its descendent Scopes (More on this later).
Container, which is the existing API for DI container, is the just the root Scope.
To make things easier to review, I have broken down the PRs into few parts:

Introduce Dig.Scope type (this PR)
Add exported provide option to Provide, which will allow Provides to a Scope to be propagated to all the Scopes without having to touch the root Scope.
Add Decorate method to Dig.Scope, which lets you decorate a type provided to a Scope.
Notes on Implementation:

Each Scope struct keeps track of the following things:

constructorNodes, providers, groups, and values all track things that were provided directly to this Scope. It does not contain anything that was provided by its parent, child, or anything in its ancestry tree. This is so that we can avoid having to copy over these every single time a new Scope is created, which is expensive in terms of memory and computation.
gh still holds all the nodes in the entire dependency graph for nodes that impact this Scope. This is so that we can avoid re-computing the dependency graph every single time a graph algorithm needs to be run.
Ref: GO-939