Add Export option to Provide (#306)
This adds Export option to Provide, which specifies that the provided constructor should be made available to all the Scopes in the same scope tree.

For example:

c := New()
s1 := c.Scope("child 1")
s2 := c.Scope("child 2")
s2.Provide(func() *A { ... })
c.Invoke(func(a *A) { ... }) // errors
s1.Invoke(func(a *A) { ...}) // errors

will error out on Invoke because constructor providing *A is provided to the s2 only.

With Export option, the child can provide this to all the Scopes in effect:

c := New()
s1 := c.Scope("child 1")
s2 := c.Scope("child 2")
s2.Provide(func() *A { ... }, Export(true))
c.Invoke(func(a *A) { ... })  // works!
s1.Invoke(func(a *A) { ... }) // works!

The implementation is quite simple - if this option is set, we provide it to the root Container. That provides the constructor visibility to all the scopes in effect.