param: Search scopes bottom up, not top-down
Instead of searching for a scope that provides everything we need
starting at a paramList, search in paramSingle.

The effect of this is that when we attempt to build a value,
it will search through the list of containers until it finds a match,
and then use that container for all consecutive operations.

This lets us simplify the error handling logic by removing it
completely.