param: Search scopes during Build, support value groups (#312)
This significantly changes how we instantiate dependencies for a
constructor and adds support for value groups.

Previously, for each parameter of a constructor, we attempted to
instantiate the value in every scope until one succeeded without
dependency errors. Effectively:

    for param in constructor
        for scope in scopes_to_root
            err = scope.instantiate(param)
            if err != missing_dependencies {
                return err
            }

This made the behavior less certain because we were relying on fuzzily
matching errors.

With this change, we search the container for constructors as needed
for singular parameters (paramSingle) and for value groups
(paramGroupedSlice). Roughly:

    for param in constructor
        switch param {
        case paramSingle:
            for scope in scopes_to_root {
                if scope.provides(param) {
                    found = true
                }
            }
            // ...
        
        case paramGroupedSlice:
            for scope in scopes_to_root {
                providers.append(scope.providers_for(param))
            }
        
        default:
            // ...
        }

This is cleaner because it works by searching the container for what it
needs rather than trying and letting it fail.

This includes a subtle change to constructorNode.Call:
previously, when a constructor ran, its results were stored in the
containerStore that it used to instantiate its dependencies.
Now, results are stored in the containerStore (Scope) to which that
constructor was provided. This works because the leaf param types
(paramSingle and paramGroupedSlice) will search the scope tree to find
the value.