Add Scope.Decorate (#313)
This adds Scope.Decorate which lets the user decorate an existing provider for a given type by overriding that with another provider.

For example,

c := dig.New()

c.Provide(func() *Logger {
  return Logger {
    Name: "Default",
  }
})

c.Decorate(func(l *Logger) *Logger {
  return Logger {
    Name: "Decorated",
  }
})
the code snippet above shows a provider that injects a *Logger type into the top-level scope. Then it adds a decorator which takes in the *Logger type and replaces its name with something else.

Once that has been done, the following Invoke will get the decorated name:

dig.Invoke(func(l *Logger) {
  l.Log(l.Name) // will log "Decorated"
})
Decorations are limited to its scope only. i.e. a child's decorator does not affect the parent scope or any of the ancestor scopes. A parent's decorator does affect the child and the descendant scopes.

One limitation: a Scope can only decorate a type once. It is possible, however, to create a child scope and then decorate once more in the child scope. In such case, the parent's decorator will be executed prior to the child's decorator.

In terms of implementation, an additional set of types were added. Namely:

decorator: This is an interface that abstract a decorator that was provided to a Scope.
decoratorNode: This implements the decorator interface.
DecorateOption: Functional options for Scope.Decorate. Unimplemented.
In addition, Scope also has new fields to keep track of all the decorators and decorated values/value groups that were injected in the scope directly. The way they work is quite similar to providers and values/value groups list. Basically, decorators are analogous to providers (constructors), and it keeps a separated list of decorated values and value groups in addition to provider-injected values and value groups.